% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{apply_reconstruction}
\alias{apply_reconstruction}
\title{Takes series of transformations and applies them to the data}
\usage{
apply_reconstruction(tree, data)
}
\arguments{
\item{tree}{transformations}

\item{data}{data to be transformed}
}
\value{
transformed data
}
\description{
Takes series of transformations and applies them to the data
}
