% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{assign_grids}
\alias{assign_grids}
\title{Assign panels to the grid}
\usage{
assign_grids(ROW, COL, width, height)
}
\arguments{
\item{ROW}{number of rows}

\item{COL}{number of columns}

\item{width}{width of panel}

\item{height}{height of panel}
}
\description{
Assign panels to the grid
}
