% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\name{axesServer}
\alias{axesServer}
\title{Server for axes module}
\usage{
axesServer(id, pipeline_variables)
}
\arguments{
\item{id}{module id}

\item{pipeline_variables}{R6 object containing important common variables}
}
\description{
Server for axes module
}
