% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonbutton.R
\name{comparisonbuttonServer}
\alias{comparisonbuttonServer}
\title{Server for comparison button module}
\usage{
comparisonbuttonServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for examine module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for comparison button module
}
