% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonbutton.R
\name{comparisonbuttonUI}
\alias{comparisonbuttonUI}
\title{UI for comparison button module}
\usage{
comparisonbuttonUI(id)
}
\arguments{
\item{id}{id for module}
}
\description{
UI for comparison button module
}
