% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine.R
\name{examineServer}
\alias{examineServer}
\title{Server for examine module}
\usage{
examineServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for examine module
}
