% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_from_paper}
\alias{example_data_from_paper}
\title{Simulated data used to generate figures in accompanying manuscript}
\format{
\subsection{\code{example_data_from_paper}}{

A data frame with 5,100 rows and 6 columns:
\describe{
\item{seconds}{x variable (e.g. time)}
\item{series}{time series to which the data belong}
\item{identity}{group to which the time series belongs}
\item{y}{y value}
\item{identity_16, identity_random}{group to which time series belongs
under slight corruption and random permutation}
...
}
}
}
\source{
simulated
}
\usage{
example_data_from_paper
}
\description{
Data intended to show specific issues with currently available
visualization tools
}
\keyword{datasets}
