% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_with_row_col}
\alias{example_data_with_row_col}
\title{Simulated data with row and column information}
\format{
\subsection{\code{example_data_with_row_col}}{

A data frame with 20,000 rows and 3 columns:
\describe{
\item{x}{x variable (e.g. time)}
\item{y}{y value}
\item{yprime}{alternative y value}
\item{z}{panel identifier}
\item{rows}{row location for panel}
\item{cols}{column location for panel}
...
}
}

\subsection{\code{running_data}}{

A data frame with 194314 rows and 48 columns:
\describe{
\item{run_id}{}
\item{limb}{}
\item{timestamp}{}
\item{step}{}
\item{stide_pace}{}
\item{step_rate}{}
\item{stride_length}{}
\item{contact_time}{}
\item{flight_ratio}{}
\item{power}{}
\item{shock}{}
\item{impact_gs}{}
\item{braking_gs}{}
\item{footstrike_type}{}
\item{pronation_excursion_fs_mp}{}
\item{max_pronation_velocity}{}
\item{grf_vert_peak}{}
\item{contact_ratio}{}
\item{flight_time}{}
\item{stride_angle}{}
\item{kleg}{}
\item{kvert}{}
\item{grf_vert}{}
\item{grf_horiz}{}
\item{swing_force_rate}{}
\item{total_force_rate}{}
\item{time_max_fs}{}
\item{time_fs_mpv}{}
\item{time_mpv_mp}{}
\item{time_mp_to}{}
\item{time_to_min}{}
\item{time_swing}{}
\item{medial_lateral_gs}{}
\item{braking_gs_amp}{}
\item{impact_gs_amp}{}
\item{vertical_speed}{}
\item{elevation_gain}{}
\item{swing_excursion}{}
\item{yaw_excursion_fs_mp}{}
\item{yaw_excursion_mp_to}{}
\item{yaw_excursion_swing}{}
\item{stance_velocity_max_fs_mp}{}
\item{stance_velocity_max_mp_to}{}
\item{temperature}{}
\item{step_length}{}
\item{pronation_excursion_mp_to}{}
\item{stance_excursion_fs_mp}{}
\item{stance_excursion_mp_to}{}
...
}
}
}
\source{
simulated

simulated
}
\usage{
example_data_with_row_col

example_data_with_row_col
}
\description{
simulated from a normal distribution-- no signal

simulated from a normal distribution-- no signal
}
\keyword{datasets}
