% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{excludeServer}
\alias{excludeServer}
\title{Server for exclude module}
\usage{
excludeServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for exclude module
}
