% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{execute_at_next_input}
\alias{execute_at_next_input}
\title{Not used?}
\usage{
execute_at_next_input(expr, session = getDefaultReactiveDomain())
}
\arguments{
\item{expr}{expression to execute}

\item{session}{session}
}
\description{
Not used?
}
