% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterbutton.R
\name{filterbuttonServer}
\alias{filterbuttonServer}
\title{Server for filter button module}
\usage{
filterbuttonServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for examine module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for filter button module
}
