% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Obtain series of ancestors to current node}
\usage{
get_ancestors(index, data)
}
\arguments{
\item{index}{current node index}

\item{data}{dataset}
}
\value{
vector of ancestors
}
\description{
Obtain series of ancestors to current node
}
