% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{get_centers_dist}
\alias{get_centers_dist}
\title{Center distance function for panel scatterplot}
\usage{
get_centers_dist(centers, width, height)
}
\arguments{
\item{centers}{center of panels}

\item{width}{width of panels}

\item{height}{height of panels}
}
\description{
Center distance function for panel scatterplot
}
