% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{get_centers_with_bounds}
\alias{get_centers_with_bounds}
\title{Obtain centers of panels}
\usage{
get_centers_with_bounds(
  data,
  scores,
  width_factor = 1,
  height_factor = 1,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{panel data}

\item{scores}{not needed?}

\item{width_factor}{width multiplier}

\item{height_factor}{height multiplier}

\item{width}{width of panel}

\item{height}{height of panel}
}
\description{
Obtain centers of panels
}
