% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{get_comparison_scores}
\alias{get_comparison_scores}
\title{Function to calculate comparison scores}
\usage{
get_comparison_scores(df, panel, fn)
}
\arguments{
\item{df}{data}

\item{panel}{panel chosen for comparison}

\item{fn}{function to perform comparison}
}
\value{
data frame with scores
}
\description{
Function to calculate comparison scores
}
