% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{get_parent_chain}
\alias{get_parent_chain}
\title{Function that returns identity of the parent node}
\usage{
get_parent_chain(tree, node)
}
\arguments{
\item{tree}{transform tree}

\item{node}{current node}
}
\value{
parent node
}
\description{
Function that returns identity of the parent node
}
