% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{get_plot_vals}
\alias{get_plot_vals}
\title{transforms data to grid formation for plotting}
\usage{
get_plot_vals(data, x, y, key)
}
\arguments{
\item{data}{data requiring transformation}

\item{x}{variable mapped to x axis}

\item{y}{variable mapped to y axis}

\item{key}{variable associated with the panel}
}
\description{
transforms data to grid formation for plotting
}
