% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{get_reconstruction_chain}
\alias{get_reconstruction_chain}
\title{Returns a series of transformations to get to current node}
\usage{
get_reconstruction_chain(tree, parent_chain)
}
\arguments{
\item{tree}{transform tree}

\item{parent_chain}{chain of nodes leading to current node}
}
\value{
series of transform functions
}
\description{
Returns a series of transformations to get to current node
}
