% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{get_the_grid}
\alias{get_the_grid}
\title{Establishes axes locations for panels and selection rectangles}
\usage{
get_the_grid(
  x,
  maxcol = 30,
  width = 0.8,
  height = 0.8,
  setup_grid = NA,
  force_grid = FALSE
)
}
\arguments{
\item{x}{list of panels}

\item{maxcol}{maximum number of columns for panel grid plot}

\item{width}{panel dimension}

\item{height}{panel dimension}

\item{setup_grid}{pre existing or user-supplied grid}

\item{force_grid}{is there a specific format for the
grid (e.g. number of columns)}
}
\description{
Establishes axes locations for panels and selection rectangles
}
