% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{get_value_scores}
\alias{get_value_scores}
\title{Compute scores from user supplied function}
\usage{
get_value_scores(data, fn1, fn2 = NULL)
}
\arguments{
\item{data}{data set}

\item{fn1}{function to calculate score}

\item{fn2}{optional function if calculating scores for scatterplot}
}
\value{
data frame with scores
}
\description{
Compute scores from user supplied function
}
