% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{labelServer}
\alias{labelServer}
\title{Server for label button module}
\usage{
labelServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for examine module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for label button module
}
