% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{min_diff}
\alias{min_diff}
\title{minimum difference between successive entries in a vector}
\usage{
min_diff(x)
}
\arguments{
\item{x}{vector}
}
\description{
minimum difference between successive entries in a vector
}
