% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newComparisonFunction.R
\name{newComparisonFunctionServer}
\alias{newComparisonFunctionServer}
\title{Server for new comparison module}
\usage{
newComparisonFunctionServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for examine module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for new comparison module
}
