% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newComparisonFunction.R
\name{newComparisonFunctionUI}
\alias{newComparisonFunctionUI}
\title{UI for new comparison module}
\usage{
newComparisonFunctionUI(id)
}
\arguments{
\item{id}{id for module}
}
\description{
UI for new comparison module
}
