% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFilterFunction.R
\name{newFilterFunctionServer}
\alias{newFilterFunctionServer}
\title{Server for new filter function module}
\usage{
newFilterFunctionServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for examine module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for new filter function module
}
