% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFilterFunction.R
\name{newFilterFunctionUI}
\alias{newFilterFunctionUI}
\title{UI for new filter function module}
\usage{
newFilterFunctionUI(id)
}
\arguments{
\item{id}{id for module}
}
\description{
UI for new filter function module
}
