% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newTransformFunction.R
\name{newTransformFunctionServer}
\alias{newTransformFunctionServer}
\title{Server for new transform function module}
\usage{
newTransformFunctionServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for examine module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for new transform function module
}
