% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newTransformFunction.R
\name{newTransformFunctionUI}
\alias{newTransformFunctionUI}
\title{UI for new transform function module}
\usage{
newTransformFunctionUI(id)
}
\arguments{
\item{id}{id for module}
}
\description{
UI for new transform function module
}
