% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{normalise}
\alias{normalise}
\title{function to normalize a vector}
\usage{
normalise(x, xmin = NULL, xmax = NULL)
}
\arguments{
\item{x}{data vector}

\item{xmin}{min supplied if min should not be derived from the vector}

\item{xmax}{max supplied if max should not be derived from the vector}
}
\description{
function to normalize a vector
}
