% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{panelsServer}
\alias{panelsServer}
\title{Server for panels module}
\usage{
panelsServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for examine module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for panels module
}
