% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placePanels.R
\name{placePanelsServer}
\alias{placePanelsServer}
\title{Server for panel placement module}
\usage{
placePanelsServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for examine module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for panel placement module
}
