% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placerbutton.R
\name{placerbuttonServer}
\alias{placerbuttonServer}
\title{Server for placement button module}
\usage{
placerbuttonServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for examine module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for placement button module
}
