% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placerbutton.R
\name{placerbuttonUI}
\alias{placerbuttonUI}
\title{UI for placement button module}
\usage{
placerbuttonUI(id, button)
}
\arguments{
\item{id}{id for module}

\item{button}{name to appear in text of button}
}
\description{
UI for placement button module
}
