% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotECDF.R
\name{plotECDFServer}
\alias{plotECDFServer}
\title{Server for ecdf module}
\usage{
plotECDFServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for ecdf module
}
