% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ppa}
\alias{ppa}
\title{Call the ppa app}
\usage{
ppa(
  mydata,
  x,
  y,
  panel,
  plottype,
  rowgroup,
  colgroup,
  randomize = NA,
  grid = NULL,
  session_name = "ppa",
  ...
)
}
\arguments{
\item{mydata}{(required) The dataset containing the data being plotted}

\item{x}{(required) variable name to be mapped to the x axis}

\item{y}{(required) variable name to be mapped to the y axis}

\item{panel}{(required) variable that indicates the specific panel to which
data belong}

\item{plottype}{currently either 'line' (default) or 'scatter' which defines
the plot type}

\item{rowgroup}{variable that groups panels by row if applicable}

\item{colgroup}{variable that groups panels by column if applicable}

\item{randomize}{currently unused}

\item{grid}{tibble with columns ROW, COL, and the 'panel' variable which
allows for a customized grid}

\item{session_name}{name for the session}

\item{...}{additional variables to be passed on}
}
\description{
Call the ppa app
}
