% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management_utils.R
\name{save_vars}
\alias{save_vars}
\title{save variables that will be useful for future analyses}
\usage{
save_vars(pipe_vars, session_name)
}
\arguments{
\item{pipe_vars}{R6 object containing global variables}

\item{session_name}{session identifier}
}
\description{
save variables that will be useful for future analyses
}
