% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreByComparison.R
\name{scoreByComparisonServer}
\alias{scoreByComparisonServer}
\title{Server for score by comparison module}
\usage{
scoreByComparisonServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for score by comparison module
}
