% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreByComparison.R
\name{scoreByComparisonUI}
\alias{scoreByComparisonUI}
\title{UI for score by comparison module}
\usage{
scoreByComparisonUI(id)
}
\arguments{
\item{id}{id for module}
}
\description{
UI for score by comparison module
}
