% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreByFunction.R
\name{scoreByFunctionServer}
\alias{scoreByFunctionServer}
\title{Server for score by function module}
\usage{
scoreByFunctionServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for score by function module
}
