% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreByFunction.R
\name{scoreByFunctionUI}
\alias{scoreByFunctionUI}
\title{UI for score by function module}
\usage{
scoreByFunctionUI(id)
}
\arguments{
\item{id}{id for module}
}
\description{
UI for score by function module
}
