% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{setup_grid_groups}
\alias{setup_grid_groups}
\title{Place panels in grid}
\usage{
setup_grid_groups(x, maxcol, force_grid = FALSE)
}
\arguments{
\item{x}{vector of panels}

\item{maxcol}{maximal columns allowed}

\item{force_grid}{TRUE if number of columns should not be derived from the number of panels}
}
\description{
Place panels in grid
}
