% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management_utils.R
\name{shuffle_col}
\alias{shuffle_col}
\title{Shuffles placement grid by column}
\usage{
shuffle_col(grid)
}
\arguments{
\item{grid}{tibble with ROW, COL, and panel}
}
\value{
grid with columns permuted from input grid
}
\description{
Shuffles placement grid by column
}
