% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management_utils.R
\name{shuffle_panel}
\alias{shuffle_panel}
\title{Shuffles placement grid by panel ignoring row and column
Note that this function does break dependence on both row and column}
\usage{
shuffle_panel(grid)
}
\arguments{
\item{grid}{tibble with ROW, COL, and panel}
}
\value{
grid with panels shuffled, ignoring row and column assignments.
}
\description{
Shuffles placement grid by panel ignoring row and column
Note that this function does break dependence on both row and column
}
