% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management_utils.R
\name{shuffle_row}
\alias{shuffle_row}
\title{Shuffles placement grid by row}
\usage{
shuffle_row(grid)
}
\arguments{
\item{grid}{tibble with ROW, COL, and panel}
}
\value{
grid with rows permuted from input grid
}
\description{
Shuffles placement grid by row
}
