% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shufflePanel.R
\name{shufflePanelServer}
\alias{shufflePanelServer}
\title{Server for shuffle panel module}
\usage{
shufflePanelServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for shuffle panel module
}
