% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{test_comparison_scores_numeric}
\alias{test_comparison_scores_numeric}
\title{Utility function to help test output type from user supplied comparison}
\usage{
test_comparison_scores_numeric(df, panel, fn)
}
\arguments{
\item{df}{data}

\item{panel}{panel chosen for comparison}

\item{fn}{function to perform comparison}
}
\value{
boolean
}
\description{
Utility function to help test output type from user supplied comparison
}
