% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{test_value_scores_numeric}
\alias{test_value_scores_numeric}
\title{Utility function to help test output type from user supplied score function}
\usage{
test_value_scores_numeric(data, fn1, fn2 = NULL)
}
\arguments{
\item{data}{data set}

\item{fn1}{function to calculate score}

\item{fn2}{optional function if calculating scores for scatterplot}
}
\value{
boolean
}
\description{
Utility function to help test output type from user supplied score function
}
