% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utils.R
\name{transform_data}
\alias{transform_data}
\title{Function to transform data according to user specified function}
\usage{
transform_data(data, transformfn)
}
\arguments{
\item{data}{dataset}

\item{transformfn}{transform function}
}
\value{
tibble with transformed data
}
\description{
Function to transform data according to user specified function
}
