% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformbutton.R
\name{transformbuttonServer}
\alias{transformbuttonServer}
\title{Server for transform button module}
\usage{
transformbuttonServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for examine module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for transform button module
}
