% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformbutton.R
\name{transformbuttonUI}
\alias{transformbuttonUI}
\title{UI for transform button module}
\usage{
transformbuttonUI(id)
}
\arguments{
\item{id}{id for module}
}
\description{
UI for transform button module
}
