% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transformServer}
\alias{transformServer}
\title{Server for transform module}
\usage{
transformServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for transform module
}
