% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_tracker.R
\name{transformTrackerServer}
\alias{transformTrackerServer}
\title{Server for transform tracker module}
\usage{
transformTrackerServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for transform tracker module
}
