% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_tracker.R
\name{transformTrackerUI}
\alias{transformTrackerUI}
\title{UI for transform tracker module}
\usage{
transformTrackerUI(id)
}
\arguments{
\item{id}{id for module}
}
\description{
UI for transform tracker module
}
