% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unselectAll.R
\name{unselectAllServer}
\alias{unselectAllServer}
\title{Server for unselect module}
\usage{
unselectAllServer(id, pipeline_variables)
}
\arguments{
\item{id}{id for module}

\item{pipeline_variables}{R6 global variable object}
}
\description{
Server for unselect module
}
