% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_with_row_col}
\alias{example_data_with_row_col}
\title{Simulated data with row and column information}
\format{
\subsection{\code{example_data_with_row_col}}{

A data frame with 20,000 rows and 3 columns:
\describe{
\item{x}{x variable (e.g. time)}
\item{y}{y value}
\item{yprime}{alternative y value}
\item{z}{panel identifier}
\item{rows}{row location for panel}
\item{cols}{column location for panel}
...
}
}
}
\source{
simulated
}
\usage{
example_data_with_row_col
}
\description{
simulated from a normal distribution-- no signal
}
\keyword{datasets}
